Move toBigNumber to a global utils file
